crowd_source/crowd_server/requirements/local.txt
-r tests.txt

django-debug-toolbar==3.6.0
ipython==8.4.0
gunicorn==20.0.4